Update OSM-ISSprOM 2019.crt
Disclaimer: This is my first attempt at contributing. Please be patient with me in case I am doing stupid things. Just a few unsystematic improvements (I hope):
Added 501.6 for some tags with highway=path because often in OSM highway=path is used for paved footways, too.
Added 501.8 for similar reasons with highway=track
Added 401 for residential=apartments because this tag can be used for apartment-block residential areas, which very often are possible to enter
Added 520 for tennis courts because it helps to see them on yellow, you can then still replace with fence manually
Added 520 for gardens and flowerbeds